Initial uniform mapping mechanism; dummy player
This commit introduces the mechanics for passing data to shader programs (like
transformation matrices, light parameters, etc.) and for demo purposes, a simple
player entity represented by a rotating triangle is implemented.

On each frame, player's scene node transformation is updated with a rotation
matrix, which is then sent to the shader as part of its parameters by the
GeometryNode.