Refactor shader parameters access and rendering
This commit changes the whole way in which the rendering occurs: scene
nodes *SCHEDULE* rendering operations, which are then sorted for better
efficiency and executed in order by the renderer. Actual shader
parameters are not set anymore to the shader in the single nodes, but
are just computed there, the submission is performed by the renderer.

Also, shader uniform blocks are now used for lighting parameters.