Move OnPlayerMove handler to gamestate
It was previously in the MovementPlanner, it is now a method of the
gamestate. It calls the movement planner, after having told to the Player
to put himself in WaitForPath action, so that the Player initiates
the movement when it receives the SetPath (called from OnPathReady).

This mechanism is the same that will be implemented and used for Build
and Attack actions.